requirements.txt
django
django-rest-swagger
djangorestframework-jwt
drf-yasg
asgiref == 3.6.0
charset-normalizer==3.0.1
certifi==2022.12.7
coreapi ==2.3.3
coreschema ==0.0.4
Django==4.1.5
djangorestframework == 3.14.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt== 5.2.2
djongo ==1.3.6
dnspython==2.3.0
drf-yasg== 1.21.4
idna==3.4
inflection== 0.5.1
itypes== 1.2.0
Jinja2==3.1.2
openapi-codec ==1.3.2
MarkupSafe==2.1.2
packaging==23.0
pycryptodomex==3.10.1
pip==22.3.1
PyJWT==1.7.1
pymongo== 3.12.3
pytz== 2022.7.1
requests ==2.28.2
ruamel.yaml== 0.17.21
setuptools ==65.6.3
simplejson== 3.18.1
sqlparse == 0.2.4
tzdata ==2022.7
uritemplate == 4.1.1
urllib3 ==1.26.14
wheel== 0.38.4